Extract Method
In the Board class the update method (related to the observer pattern)
has too many things to do with the ChessData observable methods and
contains too many attributes and parameters in it. So I create a new
method named squarePosition to interacts with these instances’ method,
hide the detail to another method
Deleted the unrelated code in ConnectionBridge class which has been
proven fail to fulfill the original function desired by the developers.
So I deleted the code due to the efficiency issue.